Reworking of structure progress save
