+++ public/components/IonAssetButton.scss
@@ -1,0 +1,14 @@
+.ion-btn {
+	.fa-cesium {
+		margin-right: 0.5em;
+	}
+
+	.caret {
+		margin-left: 1em;
+	}
+}
+
+.ion-dropdowns .dropdown {
+	float: none;
+	margin-right: 4px;
+}
