+++ templates/load_buttons.js
@@ -1,8 +1,8 @@
 PluginsAPI.Dashboard.addTaskActionButton(
-	["${app_name}/build/UploadButton.js"],
-	function(args, UploadButton) {
+	["${app_name}/build/TaskView.js"],
+	function(args, TaskView) {
 		console.log(args);
-		return React.createElement(UploadButton, {
+		return React.createElement(TaskView, {
 			task: args.task,
 			token: "${token}",
 			apiURL: "${api_url}",
